# nav-sandhu-files/explorer/app/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Hyperledger Explorer REST API Swagger",
        "description": "Rest API for fabric",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Hyperledger Team"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080"
        },
        {
            "url": "https://localhost:8080"
        }
    ],
...
    "paths": {
...
        "/auth/login": {
            "post": {
                "tags": ["Login"],
                "summary": "Login",
                "operationId": "Login",
                "requestBody": {
                    "description": "Login",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/login"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {}
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {}
                    }
                }
            }
        },
...
        "/auth/networklist": {
            "get": {
                "tags": ["Login"],
                "summary": "Query to get networklist",
                "operationId": "getNetworkList",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {}
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "login": {
                "type": "object",
                "properties": {
                    "user": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "network": {
                        "type": "string"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}